+++ src/main/java/com/huotu/ymr/mallrepository/MallUserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by lgh on 2015/12/8.
  */
-public interface MallUserRepository extends JpaRepository<MallUser, Integer>, ClassicsRepository<MallUser> {
+public interface MallUserRepository extends JpaRepository<MallUser, Long>, ClassicsRepository<MallUser> {
 //    MallUser findByMerchant(MallMerchant merchant);
 
 }
